pyproject.toml
[build-system]
requires = ["flit_core >= 3.4"]
build-backend = "flit_core.buildapi"

[project]
dynamic = ["version", "description"]
name = "pyrancid"
authors = [
    {name='Jason Hall', email='jdhall75@gmail.com'}
]
maintainers = [
  {name = "Jason Hall", email = "jdhall75@gmail.com"}
]
readme =  "README.md"
license = {file = "LICENSE"}
keywords = ["network", "backups", "version control", "configuration backup"]
dependencies = ['click','scrapli','celery','ruamel.yaml']
requires-python = ">= 3.10"

[project.scripts]
pyrancid = "pyrancid.cli:main"

[project.urls]
Repository = "git@github.com:jdhall75/pyrancid.git"
